+++ pdata/internal/common.go
@@ -172,7 +172,7 @@ func newValueFromRaw(iv interface{}) Value {
 		return mv
 	case []interface{}:
 		av := NewValueSlice()
-		newSliceFromRaw(tv).CopyTo(av.SliceVal())
+		NewSliceFromRaw(tv).CopyTo(av.SliceVal())
 		return av
 	default:
 		return NewValueString(fmt.Sprintf("<Invalid value type %T>", tv))
@@ -262,7 +262,16 @@ func (v Value) SliceVal() Slice {
 // If the Type() is not ValueTypeBytes then returns false.
 // Calling this function on zero-initialized Value will cause a panic.
 // Modifying the returned []byte in-place is forbidden.
+// Deprecated: [0.51.0] Use MBytesVal instead.
 func (v Value) BytesVal() []byte {
+	return v.orig.GetBytesValue()
+}
+
+// MBytesVal returns the []byte value associated with this Value.
+// If the Type() is not ValueTypeBytes then returns false.
+// Calling this function on zero-initialized Value will cause a panic.
+// Modifying the returned []byte in-place is forbidden.
+func (v Value) MBytesVal() []byte {
 	return v.orig.GetBytesValue()
 }
 
@@ -299,7 +308,17 @@ func (v Value) SetBoolVal(bv bool) {
 // Calling this function on zero-initialized Value will cause a panic.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use SetMBytesVal instead.
 func (v Value) SetBytesVal(bv []byte) {
+	v.orig.Value = &otlpcommon.AnyValue_BytesValue{BytesValue: bv}
+}
+
+// SetMBytesVal replaces the []byte value associated with this Value,
+// it also changes the type to be ValueTypeBytes.
+// Calling this function on zero-initialized Value will cause a panic.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (v Value) SetMBytesVal(bv []byte) {
 	v.orig.Value = &otlpcommon.AnyValue_BytesValue{BytesValue: bv}
 }
 
@@ -448,7 +467,7 @@ func (v Value) AsString() string {
 		return base64.StdEncoding.EncodeToString(v.BytesVal())
 
 	case ValueTypeSlice:
-		jsonStr, _ := json.Marshal(v.SliceVal().asRaw())
+		jsonStr, _ := json.Marshal(v.SliceVal().AsRaw())
 		return string(jsonStr)
 
 	default:
@@ -504,7 +523,7 @@ func (v Value) asRaw() interface{} {
 	case ValueTypeMap:
 		return v.MapVal().AsRaw()
 	case ValueTypeSlice:
-		return v.SliceVal().asRaw()
+		return v.SliceVal().AsRaw()
 	}
 	return fmt.Sprintf("<Unknown OpenTelemetry value type %q>", v.Type())
 }
@@ -619,11 +638,6 @@ func (m Map) Get(key string) (Value, bool) {
 	return Value{nil}, false
 }
 
-// Deprecated: [v0.47.0] Use Remove instead.
-func (m Map) Delete(key string) bool {
-	return m.Remove(key)
-}
-
 // Remove removes the entry associated with the key and returns true if the key
 // was present in the map, otherwise returns false.
 func (m Map) Remove(key string) bool {
@@ -714,7 +728,18 @@ func (m Map) InsertBool(k string, v bool) {
 // No action is applied to the map where the key already exists.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use InsertMBytes instead.
 func (m Map) InsertBytes(k string, v []byte) {
+	if _, existing := m.Get(k); !existing {
+		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
+	}
+}
+
+// InsertMBytes adds the []byte Value to the map when the key does not exist.
+// No action is applied to the map where the key already exists.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) InsertMBytes(k string, v []byte) {
 	if _, existing := m.Get(k); !existing {
 		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
 	}
@@ -769,9 +794,20 @@ func (m Map) UpdateBool(k string, v bool) {
 // No action is applied to the map where the key does not exist.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use UpdateMBytes instead.
 func (m Map) UpdateBytes(k string, v []byte) {
 	if av, existing := m.Get(k); existing {
-		av.SetBytesVal(v)
+		av.SetMBytesVal(v)
+	}
+}
+
+// UpdateMBytes updates an existing []byte Value with a value.
+// No action is applied to the map where the key does not exist.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) UpdateMBytes(k string, v []byte) {
+	if av, existing := m.Get(k); existing {
+		av.SetMBytesVal(v)
 	}
 }
 
@@ -840,9 +876,23 @@ func (m Map) UpsertBool(k string, v bool) {
 // updated to the map where the key already existed.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use UpsertMBytes instead.
 func (m Map) UpsertBytes(k string, v []byte) {
 	if av, existing := m.Get(k); existing {
-		av.SetBytesVal(v)
+		av.SetMBytesVal(v)
+	} else {
+		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
+	}
+}
+
+// UpsertMBytes performs the Insert or Update action. The []byte Value is
+// inserted to the map that did not originally have the key. The key/value is
+// updated to the map where the key already existed.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) UpsertMBytes(k string, v []byte) {
+	if av, existing := m.Get(k); existing {
+		av.SetMBytesVal(v)
 	} else {
 		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
 	}
@@ -919,8 +969,8 @@ func (m Map) AsRaw() map[string]interface{} {
 	return rawMap
 }
 
-// newSliceFromRaw creates a Slice with values from the given []interface{}.
-func newSliceFromRaw(rawSlice []interface{}) Slice {
+// NewSliceFromRaw creates a Slice with values from the given []interface{}.
+func NewSliceFromRaw(rawSlice []interface{}) Slice {
 	if len(rawSlice) == 0 {
 		v := []otlpcommon.AnyValue(nil)
 		return Slice{&v}
@@ -932,8 +982,8 @@ func newSliceFromRaw(rawSlice []interface{}) Slice {
 	return Slice{&origs}
 }
 
-// asRaw creates a slice out of a Slice.
-func (es Slice) asRaw() []interface{} {
+// AsRaw converts the Slice to a standard go slice.
+func (es Slice) AsRaw() []interface{} {
 	rawSlice := make([]interface{}, 0, es.Len())
 	for i := 0; i < es.Len(); i++ {
 		rawSlice = append(rawSlice, es.At(i).asRaw())