+++ CHANGELOG.md
@@ -4,15 +4,44 @@
 
 ### 🛑 Breaking changes 🛑
 
-- Remove deprecated model module, everything is available in `pdata` and `semconv`. (#5281)
-  - Old versions of the module are still available, but no new versions will be released.
-- Remove deprecated LogRecord.Name field. (#5202)
+- Remove deprecated pdata funcs/structs from v0.50.0 (#5345)
 
 ### 🚩 Deprecations 🚩
 
 ### 💡 Enhancements 💡
 
 ### 🧰 Bug fixes 🧰
+
+## v0.51.0 Beta
+
+### 🛑 Breaking changes 🛑
+
+- Remove deprecated model module, everything is available in `pdata` and `semconv`. (#5281)
+  - Old versions of the module are still available, but no new versions will be released.
+- Remove deprecated LogRecord.Name field. (#5202)
+
+### 🚩 Deprecations 🚩
+
+- In preparation of migration to immutable slices for primitive type items, the following methods are renamed (#5344)
+  - `Value.BytesVal` func is deprecated in favor of `Value.MBytesVal`.
+  - `Value.SetBytesVal` func is deprecated in favor of `Value.SetMBytesVal`.
+  - `Value.UpdateBytes` func is deprecated in favor of `Value.UpdateMBytes`.
+  - `Value.InsertBytes` func is deprecated in favor of `Value.InsertMBytes`.
+  - `Value.UpsertBytes` func is deprecated in favor of `Value.UpsertMBytes`.
+  - `<HistogramDataPoint|Buckets>.BucketCounts` funcs are deprecated in favor of
+    `<HistogramDataPoint|Buckets>.MBucketCounts`.
+  - `<HistogramDataPoint|Buckets>.SetBucketCounts` funcs are deprecated in favor of
+    `<HistogramDataPoint|Buckets>.SetMBucketCounts`.
+  - `HistogramDataPoint.ExplicitBounds` func is deprecated in favor of `HistogramDataPoint.MExplicitBounds`.
+  - `HistogramDataPoint.SetExplicitBounds` func is deprecated in favor of `HistogramDataPoint.SetMExplicitBounds`.
+
+### 💡 Enhancements 💡
+
+- `pdata`: Expose `pcommon.NewSliceFromRaw` and `pcommon.Slice.AsRaw` functions (#5311)
+
+### 🧰 Bug fixes 🧰
+
+- Fix Windows Event Logs ignoring user-specified logging options (#5298)
 
 ## v0.50.0 Beta
 
@@ -21,6 +50,7 @@
 - Remove pdata deprecated funcs from 2 versions (v0.48.0) ago. (#5219)
 - Remove non pdata deprecated funcs/structs (#5220)
 - `pmetric.Exemplar.ValueType()` now returns new type `ExemplarValueType` (#5233)
+- Remove deprecated `Delete` pdata func from (v0.47.0). (#5307)
 
 ### 🚩 Deprecations 🚩
 
@@ -40,6 +70,7 @@
 - Deprecate global `featuregate` funcs in favor of `GetRegistry` and a public `Registry` type (#5160)
 
 ### 💡 Enhancements 💡
+- Add `jsoniter` Unmarshaller (#4817)
 
 - Extend config.Map.Unmarshal hook to check map key string to any TextUnmarshaler not only ComponentID (#5244)
 - Collector will no longer print error with stack trace when the collector is shutdown due to a context cancel. (#5258)